Test commit.

+++ LiteFx.Bases/IIdentificador.cs
@@ -5,8 +5,11 @@ using System.Text;
 
 namespace LiteFx.Bases
 {
+    /// <summary>
+    /// To be used when a object needs to bee unique identified.
+    /// </summary>
     public interface IIdentificador
     {
-        decimal Identificador { get; set; }
+        long Identificador { get; set; }
     }
 }
